+++ src/app/admin/users/user-list/user-list-item/user-list-item.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input } from '@angular/core'
+import { Component, Input, OnInit } from '@angular/core'
 import { Router } from '@angular/router'
 import { Store } from '@ngrx/store'
 import {
@@ -11,6 +11,7 @@ import { AdminService } from 'admin/admin.service'
 import { AppState } from 'models/app-state.model'
 import { ImpersonationService } from '../../../../impersonation/impersonation.service'
 import { SessionService } from '../../../../session.service'
+import { HandleSubscription } from 'common/handle-subscription'
 import { CODE, CRYPTO } from 'common/utilities/consts'
 import { User } from 'models/user.model'
 import { ShowDialogOnError } from 'store/common/common.actions'
@@ -21,6 +22,7 @@ import {
   UserConfirmResponseDialogComponent
 } from 'common/dialog/user-confirm-response-dialog/user-confirm-response-dialog.component'
 import { BanUser, DeleteUser, UnbanUser } from 'store/admin/admin.actions'
+import { ServerOptionsService } from 'common/server-options.service'
 
 
 @Component({
@@ -28,25 +30,33 @@ import { BanUser, DeleteUser, UnbanUser } from 'store/admin/admin.actions'
   templateUrl: './user-list-item.component.html',
   styleUrls: ['./user-list-item.component.scss'],
 })
-export class UserListItemComponent {
+export class UserListItemComponent extends HandleSubscription implements OnInit {
   @Input() user: UserInfo
   loggedUser: User
   faIconEmailConfirmed = faEnvelope
   faIconAdminConfirmed = faCheck
   crypto: string = CRYPTO
   code: string = CODE
+  calculateFunds: boolean
   faIconImpersonation = faUserSecret
   isSaving: boolean = false;
 
   constructor (
     private adminService: AdminService,
+    private serverOptionsService: ServerOptionsService,
     private store: Store<AppState>,
     private impersonationService: ImpersonationService,
     private sessionService: SessionService,
     private router: Router,
     private dialog: MatDialog,
   ) {
-    this.loggedUser = sessionService.getUser()
+    super()
+  }
+
+  ngOnInit(): void {
+    this.loggedUser = this.sessionService.getUser()
+    const options = this.serverOptionsService.getOptions()
+    this.calculateFunds = options.displayCurrency !== options.appCurrency
   }
 
   handleConfirmation (): void {