+++ src/app/common/components/chart/chart-settings/chart-settings.helpers.ts
@@ -7,10 +7,6 @@ import {
   WEEK_AND_MONTH_FORMAT,
   YEAR_FORMAT,
 } from 'common/utilities/consts'
-import { formatMoney } from 'common/utilities/helpers'
-import { advChartSeriesEnum, pubChartSeriesEnum } from 'models/enum/chart.enum'
-import { ChartComponent } from 'common/components/chart/chart.component'
-import { environment } from 'environments/environment'
 
 function getFormatByDaysSpan (daysSpan: number): string {
   if (daysSpan === 0) {
@@ -33,40 +29,4 @@ export const mapDatesToChartLabels = (dates: string[]): string[] => {
   const daysSpan = values[values.length - 1].diff(values[0], 'days')
   const format = getFormatByDaysSpan(daysSpan)
   return values.map(value => value.format(format))
-}
-
-export const adjustTooltipValueFormat = (value: string): string => {
-  const type = ChartComponent.seriesType
-
-  switch (type) {
-    case advChartSeriesEnum.sum:
-    case advChartSeriesEnum.sumPayment:
-    case advChartSeriesEnum.cpc:
-    case advChartSeriesEnum.cpm:
-    case pubChartSeriesEnum.sum:
-    case pubChartSeriesEnum.sumHour:
-    case pubChartSeriesEnum.rpm:
-      const val = parseInt(value)
-      return `${type}: ${val > 0 ? formatMoney(val, 2) : 0} ${environment.currencyCode}`
-    default:
-      return `${type}: ${value}`
-  }
-}
-
-export const adjustYAxesTics = (value) => {
-  const type = ChartComponent.seriesType
-
-  switch (type) {
-    case advChartSeriesEnum.sum:
-    case advChartSeriesEnum.sumPayment:
-    case advChartSeriesEnum.cpc:
-    case advChartSeriesEnum.cpm:
-    case pubChartSeriesEnum.sum:
-    case pubChartSeriesEnum.sumHour:
-    case pubChartSeriesEnum.rpm:
-      const val = parseInt(value)
-      return `${environment.currencySymbol}${val > 0 ? formatMoney(val, 2) : 0}`
-    default:
-      return `${value}`
-  }
 }