+++ src/app/models/enum/link.enum.ts
@@ -6,9 +6,17 @@ export const ADD_UNIT_DECENTRALAND_SMART = 'https://adshar.es/dclsmart'
 export const ADSHARES_UNITS = 'https://adshares.net/units'
 // Adshares wallet installation instruction
 export const ADSHARES_WALLET = 'https://adshares.net/wallet'
+// Instruction for advertiser
+export const ADVERTISER_INSTRUCTION_LINK = 'https://github.com/adshares/adserver/wiki/How-to-use-Advertising-Ecosystem#how-to-use-the-adserver--instruction-for-advertisers'
 // Decentraland builder
 export const DECENTRALAND_BUILDER = 'https://builder.decentraland.org/'
+// How to get ADS?
+export const GET_ADS_FAQ = 'https://adshar.es/getADS'
 // Metamask wallet installation instruction
 export const METAMASK_WALLET = 'https://metamask.io/'
+// Instruction for publisher
+export const PUBLISHER_INSTRUCTION_LINK = 'https://github.com/adshares/adserver/wiki/How-to-use-Advertising-Ecosystem#how-to-use-the-adserver--instruction-for-publishers'
 // wiki guide on how to add an html ad
 export const WIKI_HTML_AD = 'https://adshar.es/howtohtmlbanner'
+// withdraw ADS faq
+export const WITHDRAW_ADS_FAQ = 'https://adshar.es/withdrawADS'