+++ src/app/settings/settings-routing.module.ts
@@ -5,11 +5,15 @@ import { SettingsComponent } from './settings.component';
 import { AccountSettingsComponent } from './general-settings/account-settings/account-settings.component';
 import { BillingComponent } from './billing/billing.component';
 import { ReportsListWrapperComponent } from 'settings/reports-list-wrapper/reports-list-wrapper.component';
+import { ServerOptionsResolver } from 'common/resolvers/server-options.resolver'
 
 const settingsRoutes: Routes = [
   {
     path: 'settings',
     component: SettingsComponent,
+    resolve: {
+      options: ServerOptionsResolver,
+    },
     children: [
       {path: 'general', component: AccountSettingsComponent},
       {