+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
+- Dynamically load currency and other AdServer's options
 - Decrease index replace delay
 
 ## [1.18.7] - 2022-07-29